organize physics concerns a bit
