+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
-export { useInitialConditions } from "./use-initial-conditions"
 export { useEntropy } from "./use-entropy"
+export { useInitialConditions } from "./use-initial-conditions"
+export { usePhysics } from "./use-physics"
